+++ modules/Custom/MyReport.php
@@ -1,6 +1,6 @@
 <?php
 
-DrawHeader(_(ProgramTitle()));
+DrawHeader( ProgramTitle() );
 
 if ( ! $_REQUEST['modfunc'] )
 {
@@ -106,7 +106,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 		foreach ( (array) $custom_fields_RET as $field )
 		{
-			$LO_columns += [ 'CUSTOM_' . $field['ID'] => ParseMLField( $field['TITLE'] ) ];
+			$LO_columns += [ 'CUSTOM_' . $field['ID'] => _makeFieldTitle( $field['TITLE'] ) ];
 		}
 
 		$LO_columns += [
@@ -136,7 +136,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 			$extra['functions'][ $field_key ] = makeFieldTypeFunction( $field['TYPE'] );
 
-			$LO_columns[ $field_key ] = ParseMLField( $field['TITLE'] );
+			$LO_columns[ $field_key ] = _makeFieldTitle( $field['TITLE'] );
 		}
 
 		$LO_columns += [ 'PERSON_NAME' => _( 'Person Name' ) ];
@@ -147,7 +147,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 			$extra['functions'][ $field_key ] = makeFieldTypeFunction( $field['TYPE'] );
 
-			$LO_columns[ $field_key ] = ParseMLField( $field['TITLE'] );
+			$LO_columns[ $field_key ] = _makeFieldTitle( $field['TITLE'] );
 		}
 
 		$students_RET = GetStuList( $extra );
@@ -216,4 +216,30 @@ function _makeTV( $value, $column )
 	$i = mb_substr( $column, 6 );
 
 	return isset( $person_RET[ $i ][ $tv ] ) ? $person_RET[ $i ][ $tv ] : null;
+}
+
+/**
+ * Make (Student, Contact, Address) Field Title
+ * Parse Multi-lingual value
+ * Truncate column title to 36 chars if > 36 chars
+ *
+ * Local function.
+ *
+ * @since 11.0
+ *
+ * @param  string $value  Title value.
+ * @param  string $column Column. Defaults to ''.
+ *
+ * @return string         Title truncated to 36 chars.
+ */
+function _makeFieldTitle( $value, $column = '' )
+{
+	$field_title = ParseMLField( $value );
+
+	// Truncate file name if > 36 chars.
+	$field_title_display = mb_strlen( $field_title ) <= 36 ?
+		$field_title :
+		'<span title="' . AttrEscape( $field_title ) . '">' . mb_substr( $field_title, 0, 33 ) . '...</span>';
+
+	return $field_title_display;
 }