+++ CHANGES.md
@@ -14,6 +14,7 @@ Changes in 11.0
 - Remove UserCoursePeriodSchoolPeriod() & UserPeriod() functions in Current.php
 - SQL access_log: deprecate LOGIN_TIME column, use CREATED_AT instead in index.php, AccessLog.php, rosariosis.sql & rosariosis_mysql.sql
 - SQL portal_notes & portal_polls: deprecate PUBLISHED_DATE column, use CREATED_AT instead in PortalNotes.php, PortalPolls.php, Portal.php, rosariosis.sql & rosariosis_mysql.sql
+- Truncate column title to 36 chars if > 36 chars in MyReport.php
 
 Changes in 10.9.3
 -----------------